refactoring: App content moved to todoList element, new elements Contacts and About me
