+++ src/components/About/About.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const About = () => <div>About</div>;
+
+export default About;
