+++ src/components/CheckboxLabels/CheckboxLabels.js
@@ -1,10 +1,10 @@
-import React from ' react';
-import PropTypes from ' prop-types';
-import FormGroup from ' @material-ui/core/FormGroup';
-import FormControlLabel from ' @material-ui/core/FormControlLabel';
-import Checkbox from ' @material-ui/core/Checkbox';
-import styles from ' ./CheckboxLabels.module.css';
-import classnames from ' classnames';
+import React from 'react';
+import PropTypes from 'prop-types';
+import FormGroup from '@material-ui/core/FormGroup';
+import FormControlLabel from '@material-ui/core/FormControlLabel';
+import Checkbox from '@material-ui/core/Checkbox';
+import styles from './CheckboxLabels.module.css';
+import classnames from 'classnames';
 
 class CheckboxLabels extends React.Component {
   componentDidMount() {
