+++ src/components/Contacts/Contacts.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const Contacts = () => <div>Render me, and I will render you!</div>;
+
+export default Contacts;
