+++ src/components/Footer/Footer.module.css
@@ -1,6 +1,5 @@
 .footer {
-    background-color: #c5c5c5;
-    padding: 5px 15px;
-    margin: 0 -20px;
-    border-radius: 0 0 10px 10px;
+  background-color: #c5c5c5;
+  padding: 5px 15px;
+  margin: 0 -20px;
 }
